Also relax the PySequence check when extracting fixed-sized arrays.

+++ src/types/sequence.rs
@@ -309,7 +309,7 @@ where
         }
     };
 
-    let mut v = Vec::with_capacity(seq.len().unwrap_or(0) as usize);
+    let mut v = Vec::with_capacity(seq.len().unwrap_or(0));
     for item in seq.iter()? {
         v.push(item?.extract::<T>()?);
     }
